1.0.4 - Structural overhaul
I was able to debug the issue, the wrapper <div> and <body> elements were misplaced, causing the head information to be read as part of the body. The changes reflect a structural overhaul of the code to a more functional version. The same changes will be made to all html files on the site.