Update 1.1.0.0 - Graphical overhaul + new functions
Added a banner at the top and bottom of the website. Added pictures and place-holder links at the top for social media links. Added logo place-holder image and functionality to go straight to this repository by clicking on the logo. Added a changelog and removed redundant information to improve page loading.